requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
aiosqlite==0.21.0
async-timeout==5.0.1
attrs==25.1.0
Brotli==1.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
frozenlist==1.5.0
g4f==0.4.7.2
idna==3.10
multidict==6.1.0
nest-asyncio==1.6.0
pillow==11.1.0
propcache==0.2.1
pyaes==1.6.1
pyasn1==0.6.1
pycryptodome==3.21.0
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
Telethon==1.38.1
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
